Runner/pyproject.toml
[project]
name = "agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.82.1",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
